move labels selection sheet to new file
